Refactoring the front-end code. It was a mess and still is, but it is improving.
