+++ package.json
@@ -5,7 +5,8 @@
   "scripts": {
     "start": "NODE_ENV=production nodejs app.js",
     "test": "PORT=3000 NODE_ENV=test nodejs app.js",
-    "build-front": "webpack --config webpack.config.js"
+    "build-front": "webpack --config webpack.config.js",
+    "build-front-watch": "webpack --watch --config webpack.config.js"
   },
   "keywords": [
     "waterrower",
