+++ public/js/src/map/index.js
@@ -1,0 +1,6 @@
+var livePoints = [];
+var liveMap = {};
+var liveBounds = {};
+
+
+export default { livePoints, liveMap, liveBounds }
