+++ public/js/src/utils/globals.js
@@ -1,0 +1,4 @@
+const RATION = (100 / 4.805) * 6;
+const PAGE_SIZE = 10;
+
+export default { RATION, PAGE_SIZE }
