+++ public/js/src/utils/utils.js
@@ -20,4 +20,19 @@ function getHeartRateColor(hr) {
     }
 }
 
-export default { calcWatt, sessionNameToReadable, getHeartRateColor }
+var getUrlParameter = function getUrlParameter(sParam) {
+    var sPageURL = decodeURIComponent(window.location.search.substring(1)),
+        sURLVariables = sPageURL.split('&'),
+        sParameterName,
+        i;
+
+    for (i = 0; i < sURLVariables.length; i++) {
+        sParameterName = sURLVariables[i].split('=');
+
+        if (sParameterName[0] === sParam) {
+            return sParameterName[1] === undefined ? true : sParameterName[1];
+        }
+    }
+};
+
+export default { calcWatt, sessionNameToReadable, getHeartRateColor, getUrlParameter }
